+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # Braindrop ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Made some cosmetic changes to the raindrop input dialog.
+  ([#194](https://github.com/davep/braindrop/pull/194))
+
 ## v0.13.0
 
 **Released: 2025-11-03**